Use setfiletype to allow overrides
Ansible users can choose which syntax to use for files. For example
a user could write a static inventory file - hosts - in INI or YAML
syntax [1].

`:help setfiletype` explains that `setfiletype` will:

    Set the 'filetype' option to {filetype}, but only if not done yet

This allows the user to write their own ftdetect autocommands to choose
a different file type from the default.

Without this change ftdetect auto commands from this package are read
much later in vim's startup process and therefore take precedence
over the user's own. Without this change, this behaviour is difficult to
work around, for example it is very difficult to write ftdetect
autocommands that set the hosts file syntax to YAML.

[1] http://docs.ansible.com/ansible/latest/intro_inventory.html#inventory