Add 'default' to highlight commands to enable overrides
* Add default to highlight links in ansible_hosts.vim

Adding default causes the link to be ignored if there is already a link
in place for the group. This would allow users/colorscheme owners to
override the links set in syntax/ansible_hosts.vim.

See :hi-link for more info.

* Add default to hi links in ansible.vim syntax

This only adds default to ansible specific/defined groups, and then only
groups which have not been overridden by an ansible-vim global config
flag.